fixed environment hook
In order for environment hooks to work, they have to use /bin/sh
and not bash!.

Using bash works if you are directly compiling a package.  It does not
work when it is a dependency. In that case the build system will
call setup.sh (and not setup.bash) and the bash-script won't work.